move glossary components to library and hide definition in tables, contributes to DEV-274
