+++ libs/lib-glossary/README.md
@@ -1,0 +1,7 @@
+# lib-glossary
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test lib-glossary` to execute the unit tests via [Jest](https://jestjs.io).
