+++ libs/lib-glossary/project.json
@@ -1,0 +1,9 @@
+{
+  "name": "lib-glossary",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/lib-glossary/src",
+  "projectType": "library",
+  "tags": [],
+  "// targets": "to see all targets run: nx show project lib-glossary --web",
+  "targets": {}
+}
