+++ libs/lib-glossary/src/index.ts
@@ -1,0 +1,3 @@
+// Use this file to export React client components (e.g. those with 'use client' directive) or other non-server utilities
+
+export * from './lib/lib-glossary';
