+++ libs/lib-glossary/src/lib/FilterCanonPathDropdown.tsx
