+++ libs/lib-glossary/src/ssr.ts
@@ -1,0 +1,2 @@
+export * from './lib/DetailPage';
+export * from './lib/LandingPage';
