+++ tsconfig.base.json
@@ -22,9 +22,11 @@
       "@lib-canon": ["libs/lib-canon/src/index.ts"],
       "@lib-canon/ssr": ["libs/lib-canon/src/ssr.ts"],
       "@lib-editing": ["libs/lib-editing/src/index.ts"],
-      "@lib-utils": ["libs/lib-utils/src/index.ts"],
       "@lib-user": ["libs/lib-user/src/index.ts"],
-      "@lib-user/ssr": ["libs/lib-user/src/ssr.ts"]
+      "@lib-user/ssr": ["libs/lib-user/src/ssr.ts"],
+      "@lib-utils": ["libs/lib-utils/src/index.ts"],
+      "@lib-glossary": ["libs/lib-glossary/src/index.ts"],
+      "@lib-glossary/ssr": ["libs/lib-glossary/src/ssr.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
